refactor: memberRepository 서브 패키지 제거에 따른 위치 변경
